# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CompanyOS Core API",
    "description": "Phase 1.3 \u2014 dashboard BFF + workspace + auth + hello.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/finance/credit-notes": {
      "post": {
        "tags": [
          "finance-credit-notes"
        ],
        "summary": "POST /api/v1/finance/credit-notes \u2014 create and issue immediately.",
        "operationId": "create_credit_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCreditNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditNoteDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCreditNoteRequest": {
        "type": "object",
        "required": [
          "invoice_id",
          "lines"
        ],
        "properties": {
          "invoice_id": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineInput"
            }
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreditNoteDto": {
        "type": "object",
        "required": [
          "id",
          "invoice_id",
          "customer_id",
          "credit_number",
          "currency",
          "subtotal_minor",
          "tax_minor",
          "total_minor",
          "issued_at"
        ],
        "properties": {
          "credit_number": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoice_id": {
            "type": "string"
          },
          "issued_at": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "subtotal_minor": {
            "type": "integer",
            "format": "int64"
          },
          "tax_minor": {
            "type": "integer",
            "format": "int64"
          },
          "total_minor": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "InvoiceLineInput": {
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unit_price_minor"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "discount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int64"
          },
          "tax_rate_bps": {
            "type": "integer",
            "format": "int64"
          },
          "unit_price_minor": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}